+++ src/frontend/apps/impress/src/features/docs/doc-management/components/DocIcon.tsx
@@ -8,7 +8,7 @@ import { useDocTitleUpdate } from '../hooks/useDocTitleUpdate';
 
 type DocIconProps = TextType & {
   emoji?: string | null;
-  emojiPicker?: boolean;
+  withEmojiPicker?: boolean;
   defaultIcon: React.ReactNode;
   docId?: string;
   title?: string;
@@ -17,11 +17,11 @@ type DocIconProps = TextType & {
 
 export const DocIcon = ({
   emoji,
+  withEmojiPicker = false,
   defaultIcon,
   $size = 'sm',
   $variation = '1000',
   $weight = '400',
-  emojiPicker = false,
   docId,
   title,
   onEmojiUpdate,
@@ -37,12 +37,12 @@ export const DocIcon = ({
     left: number;
   }>({ top: 0, left: 0 });
 
-  if (!emojiPicker && !emoji) {
+  if (!withEmojiPicker && !emoji) {
     return defaultIcon;
   }
 
   const toggleEmojiPicker = (e: React.MouseEvent) => {
-    if (emojiPicker) {
+    if (withEmojiPicker) {
       e.stopPropagation();
       e.preventDefault();