+++ src/frontend/apps/impress/src/features/docs/doc-management/components/index.ts
@@ -2,3 +2,4 @@ export * from './DocIcon';
 export * from './DocPage403';
 export * from './ModalRemoveDoc';
 export * from './SimpleDocItem';
+export * from './DocIcon';